Refactor dpsim class into 2 files
